Product updates, Listing and Properties creation
Product has been updated with anew state, new handleVlaue, handleChange and handleSubmit to make things more streamlined

Listing component was created to show whats being saved when the form submits and ultimately be taken to a new page.

Properties component created to contain the propertyName and propertyValue parts of state and the form. May or may not stay this way in the end.